[crater] Move Config type into sources module
